Add general ValidationException & AccessDeniedException to Api class
- (sub classes of) the new generic application level exceptions are mapped to the according status codes (400 / 403)
- Also add missing marker / stream ID to POST bodies of new job creations
- Also improved logging a bit further
